And now you can select multiple categories at once.
Auto-toggled all buttons in mapLoad just so that initially clicking to remove categories makes sense. Would be nice if we set it up so that the first-ever filter click negated all the categories except the one selected, then after that clicking any individual category would toggle it on and off.